# [QA] optimisation requetes