## Ticket

#3399

## Description
1
https://sentry.incubateur.net/organizations/betagouv/issues/127893/?environment=prod&project=61&query=is%3Aunresolved&referrer=issue-stream&statsPeriod=90d&stream_index=1

https://sentry.incubateur.net/organizations/betagouv/issues/133207/?environment=prod&project=61&query=is%3Aunresolved&referrer=issue-stream&statsPeriod=90d&stream_index=4

Ajout d'un index composé sur Suivi afin accélérer cette requête (chargement du tableau de bord) 
Durée de la requête avant/après lors de mes essaies : 
pour SA : 2500 -> 570 ms
pour RT 13 : 730 -> 480ms
pour Agent marseille : 3800 -> 540ms

2
https://sentry.incubateur.net/organizations/betagouv/issues/137000/?environment=prod&project=61&query=is%3Aunresolved&referrer=issue-stream&statsPeriod=90d&stream_index=1
Affichage de la page signalement FO (cette requête remonte récemment bien que les temps restent acceptables)

Ajout d'un index sur Signalement sur code_suivi
Test avant après : 330 -> 2 ms

3
https://sentry.incubateur.net/organizations/betagouv/issues/121522/events/2bd3936928eb41ebbdb5cee1be42eab5/
C'est la grosse requête par défaut de la liste signalement pour les SA

J'ai essayé plein de truc sur la requête et les index sans réussir à baisser a moins de 3sec
Je propose de mettre un filtre par défaut pour les SA

## Pré-requis
Se mettre sur une copie base de prod
`make load-migrations`
`make clear-pool pool=--all`

## Tests
- [ ] Regarder le temps d'execution de la requête numéro 13 du tableau de bord (bo/widget/data-kpi)
- [ ] Regarder le temps d'exécution coté fiche signalement usager